agent_with_web_search/requirements.txt
langchain
langgraph
langchain-google-genai
langchain-tavily
tavily-python
python-dotenv